@@ -28,35 +28,36 @@
     "serve": "cd tools && npm run cli:serve"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.23",
-    "@angular-devkit/build-ng-packagr": "^0.803.23",
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "^8.2.3",
-    "@angular/cli": "^8.3.23",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "^8.2.14",
+    "@angular-devkit/build-angular": "~12.2.16",
+    "@angular-devkit/core": "^12.2.16",
+    "@angular/animations": "^12.2.16",
+    "@angular/cdk": "^12.2.13",
+    "@angular/cli": "^12.2.16",
+    "@angular/common": "^12.2.16",
+    "@angular/compiler": "^12.2.16",
+    "@angular/compiler-cli": "^12.2.16",
+    "@angular/core": "^12.2.16",
+    "@angular/flex-layout": "^12.0.0-beta.35",
+    "@angular/forms": "^12.2.16",
     "@angular/http": "^7.2.16",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
+    "@angular/material": "^12.2.13",
+    "@angular/platform-browser": "^12.2.16",
+    "@angular/platform-browser-dynamic": "^12.2.16",
+    "@angular/router": "^12.2.16",
     "@compodoc/compodoc": "^1.1.7",
-    "@firestitch/clear": "^8.0.0",
-    "@firestitch/common": "^8.0.0",
-    "@firestitch/date": "^8.0.1",
-    "@firestitch/datepicker": "^8.1.2",
-    "@firestitch/drawer": "^8.0.2",
-    "@firestitch/example": "^8.0.0",
-    "@firestitch/iframe": "^8.0.0",
-    "@firestitch/label": "^8.0.0",
-    "@firestitch/menu": "^7.0.7",
-    "@firestitch/message": "^8.1.0",
-    "@firestitch/scroll-picker": "^8.0.0",
-    "codelyzer": "^5.2.1",
+    "@firestitch/clear": "^12.0.0",
+    "@firestitch/common": "^12.0.0",
+    "@firestitch/date": "^12.0.0",
+    "@firestitch/datepicker": "^12.1.0",
+    "@firestitch/drawer": "^12.0.0",
+    "@firestitch/example": "^12.0.0",
+    "@firestitch/iframe": "^12.0.0",
+    "@firestitch/label": "^12.0.0",
+    "@firestitch/menu": "^12.0.1",
+    "@firestitch/message": "^12.1.0",
+    "@firestitch/scroll-picker": "^12.0.0",
+    "@firestitch/style": "^12.0.0",
+    "codelyzer": "^6.0.0",
     "cpx": "^1.5.0",
     "date-fns": "2.9.0",
     "date-fns-tz": "^1.0.9",
@@ -64,16 +65,17 @@
     "highlight.js": "^9.18.0",
     "iscroll": "^5.2.0",
     "lodash-es": "^4.17.15",
-    "ng-packagr": "^5.7.1",
-    "ngx-toastr": "^11.3.2",
+    "ng-packagr": "^12.2.7",
+    "ngx-highlightjs": "^6.1.2",
+    "ngx-toastr": "^14.2.2",
     "npm-run-all": "^4.1.5",
+    "parse-messy-time": "^2.1.0",
     "rxjs": "^6.5.4",
-    "tsickle": "^0.38.0",
-    "tslint": "^6.0.0",
-    "typescript": "^3.5.3",
+    "tslint": "~6.1.0",
+    "typescript": "4.3.5",
     "webpack": "^4.41.5",
     "webpack-dev-server": "^3.10.1",
-    "zone.js": "^0.10.2"
+    "zone.js": "~0.11.4"
   },
   "engines": {
     "node": ">=6.0.0"
@@ -82,5 +84,7 @@
     "@firestitch/menu": ">=7.0.7",
     "lodash-es": ">=4.17.0"
   },
-  "dependencies": {}
+  "dependencies": {
+    "tslib": "^2.0.0"
+  }
 }
